backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
python-dateutil
